Implements logout functionality
